CommunicationRecords updated
