+++ frontend/src/components/pt-ot-st/patients/Patients/InfoPaciente/PatientInfoPage.jsx
@@ -9,7 +9,7 @@ import DisciplinesComponent from './DisciplinesComponent';
 import ScheduleComponent from './ScheduleComponent';
 import ExercisesComponent from './ExercisesComponent';
 import DocumentsComponent from './DocumentsComponent';
-import CommunicationRecords from '../../../../developer/patients/Patients/InfoPaciente/CommunicationRecords';
+// import CommunicationRecords from '../../../../developer/patients/Patients/InfoPaciente/CommunicationRecords'; // TEMPORARILY COMMENTED
 import LogoutAnimation from '../../../../../components/LogOut/LogOut';
 import '../../../../../styles/developer/Patients/InfoPaciente/PatientInfoPage.scss';
 
@@ -455,7 +455,10 @@ const NotesSection = ({ patient, currentCertPeriod }) => {
   
   return (
     <div className="communication-records-section">
-      <CommunicationRecords patient={patient} currentCertPeriod={currentCertPeriod} />
+      {/* <CommunicationRecords patient={patient} currentCertPeriod={currentCertPeriod} /> */}
+      <div style={{padding: '20px', textAlign: 'center', color: '#666'}}>
+        Communication Records temporarily disabled
+      </div>
     </div>
   );
 };
